--- src/app/environments/environments.ts
@@ -1,12 +1,0 @@
-export const environment = {
-  production: false,
-  firebase: {
-    apiKey: 'AIzaSyB1vM5h-Lf-Yn8hL-fQYdfEotmRn8YCPVA',
-    authDomain: 'purduechat-ryan.firebaseapp.com',
-    projectId: 'purduechat-ryan',
-    storageBucket: 'purduechat-ryan.appspot.com',
-    messagingSenderId: '572797052539',
-    appId: '1:572797052539:web:ce4b1d809c97a9f08b64de',
-    measurementId: 'G-XNPDPBD74T',
-  },
-};
