+++ src/main.ts
@@ -2,7 +2,7 @@ import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
-import { environment } from 'src/app/environments/environments';
+import { environment } from 'src/environments/environments';
 
 if (environment.production) {
   enableProdMode();
